Changes required for large DBs like VisualGenome.

+++ code/factorbase/src/main/java/ca/sfu/cs/factorbase/data/DataSet.java
@@ -5,7 +5,7 @@ package ca.sfu.cs.factorbase.data;
  */
 public class DataSet {
     private boolean isDiscrete;
-    private long[][] data;
+    private double[][] data;
     private DataSetMetaData metadata;
 
 
@@ -16,7 +16,7 @@ public class DataSet {
      * @param metadata - the associated metadata for the given {@code data}.
      * @param isDiscrete - true if the dataset only contains discrete information; otherwise false.
      */
-    public DataSet(long[][] data, DataSetMetaData metadata, boolean isDiscrete) {
+    public DataSet(double[][] data, DataSetMetaData metadata, boolean isDiscrete) {
         this.data = data;
         this.metadata = metadata;
         this.isDiscrete = isDiscrete;
@@ -28,7 +28,7 @@ public class DataSet {
      *
      * @return a 2D array representation of the dataset where each state has been converted into an indexed value.
      */
-    public long[][] getData() {
+    public double[][] getData() {
         return this.data;
     }
