+++ code/factorbase/src/main/java/ca/sfu/cs/factorbase/search/BDeuScoreOnDemand.java
@@ -65,7 +65,7 @@ public class BDeuScoreOnDemand implements DiscreteLocalScore {
             BigDecimal score = new BigDecimal((r - 1) * q * Math.log(this.structurePrior));
 
             for (List<RandomVariableAssignment> parentAssignments : meta.getStates(ct.getParentColumnIndices())) {
-                long countsSum = 0;
+                double countsSum = 0;
                 long counts;
                 for (int childStateIndex = 0; childStateIndex < r; childStateIndex++) {
                     RandomVariableAssignment childAssignment = new RandomVariableAssignment(ct.getChildColumnIndex(), childStateIndex);
